refactor(api): update product and user endpoints in index.ts files
