+++ src/api/products/index.ts
@@ -2,7 +2,7 @@ import { HTTPMethod } from '@/enums/method'
 import { Product } from '@/schemas/product.dto'
 import type { Bindings } from '@/utils/bindings'
 import { OpenAPIHono as Hono, createRoute, z } from '@hono/zod-openapi'
-import { cache } from 'hono/cache'
+import {} from 'hono/adapter'
 import Stripe from 'stripe'
 
 export const app = new Hono<{ Bindings: Bindings }>()
@@ -18,12 +18,12 @@ app.openapi(
     request: {
       query: Product.Param
     },
-    middleware: [
-      cache({
-        cacheName: 'products',
-        cacheControl: 'max-age=3600'
-      })
-    ],
+    // middleware: [
+    //   cache({
+    //     cacheName: 'products',
+    //     cacheControl: 'max-age=3600'
+    //   })
+    // ],
     responses: {
       200: {
         content: {
@@ -39,7 +39,7 @@ app.openapi(
   async (c) => {
     const param = c.req.valid('query')
     const stripe = new Stripe(c.env.STRIPE_SECRET_KEY, {
-      apiVersion: '2024-06-20',
+      apiVersion: '2025-02-24.acacia',
       typescript: true
     })
     const products = (await stripe.products.list({ limit: param.limit, active: true })).data.map((product) => Product.Data.parse(product))
